Basic Active Ragdoll Implementation
Made a test character, "Boxman." Control him with "W" and "S" to
contract or extend.
Uses Puppet Master for all physics behaviours.
Constrained his root to a 2D plane.

This is the first implementation. Next implementation will use a better
rigged character. There's some problems with the ragdoll constraints
because of the rig setup in blender. I might work on grabbing things
with the hands colliders in the next implementation.